requirements.txt
streamlit
pandas
altair
plotly
azure-storage-file-datalake
dotenv